vendor pay invoice tests
